+++ CHANGELOG.md
@@ -355,3 +355,9 @@
 
 ## 0.1.489
 * ZwapSelect added delay between fetch and on empty responses
+
+
+## 0.1.490
+* New ZwapColors: Primary500, Primary600, Primary700Dark, Primary900Dark, Neutral300Dark,Error25, Error50 
+* ZwapButtons added pressed state
+* ZwapButtonDecorations from Design System 2.0